refactor: 가독성을 위한 전체적인 이름 변경
