+++ src/main/java/lotto/domain/LottoPrize.java
@@ -2,19 +2,19 @@ package lotto.domain;
 
 public enum LottoPrize {
     FIRST_PRIZE(2000000000, "6개 일치",
-            (matchNumberCount, hasMagicNumber) -> matchNumberCount == LottoNumbers.LOTTO_NUMBER_LENGTH),
+            (matchNumberCount, hasMagicNumber) -> matchNumberCount == Lotto.LOTTO_NUMBER_LENGTH),
     SECOND_PRIZE(30000000, "5개 일치, 보너스 볼 일치",
-            (matchNumberCount, hasMagicNumber) -> matchNumberCount == LottoNumbers.LOTTO_NUMBER_LENGTH - 1
+            (matchNumberCount, hasMagicNumber) -> matchNumberCount == Lotto.LOTTO_NUMBER_LENGTH - 1
                     && hasMagicNumber),
     THIRD_PRIZE(1500000, "5개 일치",
-            (matchNumberCount, hasMagicNumber) -> matchNumberCount == LottoNumbers.LOTTO_NUMBER_LENGTH - 1
+            (matchNumberCount, hasMagicNumber) -> matchNumberCount == Lotto.LOTTO_NUMBER_LENGTH - 1
                     && !hasMagicNumber),
     FOURTH_PRIZE(50000, "4개 일치",
-            (matchNumberCount, hasMagicNumber) -> matchNumberCount == LottoNumbers.LOTTO_NUMBER_LENGTH - 2),
+            (matchNumberCount, hasMagicNumber) -> matchNumberCount == Lotto.LOTTO_NUMBER_LENGTH - 2),
     FIFTH_PRIZE(5000, "3개 일치",
-            (matchNumberCount, hasMagicNumber) -> matchNumberCount == LottoNumbers.LOTTO_NUMBER_LENGTH - 3),
+            (matchNumberCount, hasMagicNumber) -> matchNumberCount == Lotto.LOTTO_NUMBER_LENGTH - 3),
     NONE(0, "0개 일치",
-            (matchNumberCount, hasMagicNumber) -> matchNumberCount < LottoNumbers.LOTTO_NUMBER_LENGTH - 3);
+            (matchNumberCount, hasMagicNumber) -> matchNumberCount < Lotto.LOTTO_NUMBER_LENGTH - 3);
 
     private final long prizeMoney;
     private final String description;
