+++ src/main/java/lotto/utils/RandomLottoGenerator.java
@@ -5,7 +5,7 @@ import java.util.Collections;
 import java.util.List;
 import java.util.stream.Collectors;
 import java.util.stream.IntStream;
-import lotto.domain.LottoNumbers;
+import lotto.domain.Lotto;
 import lotto.domain.SingleLottoNumber;
 
 public class RandomLottoGenerator {
@@ -18,21 +18,21 @@ public class RandomLottoGenerator {
                     .mapToObj(SingleLottoNumber::new)
                     .collect(Collectors.toList());
 
-    public static List<LottoNumbers> generateLottos(int size) {
-        List<LottoNumbers> lottoNumbers = new ArrayList<>();
+    public static List<Lotto> generateLottoList(int size) {
+        List<Lotto> lottoList = new ArrayList<>();
 
         for (int i = 0; i < size; i++) {
             List<SingleLottoNumber> chosen = getRandomLottoNumbers();
-            lottoNumbers.add(new LottoNumbers(chosen));
+            lottoList.add(new Lotto(chosen));
         }
 
-        return lottoNumbers;
+        return lottoList;
     }
 
     private static List<SingleLottoNumber> getRandomLottoNumbers() {
         Collections.shuffle(lottoBalls);
         return lottoBalls.stream()
-                .limit(LottoNumbers.LOTTO_NUMBER_LENGTH)
+                .limit(Lotto.LOTTO_NUMBER_LENGTH)
                 .sorted()
                 .collect(Collectors.toList());
     }
