+++ src/test/java/lotto/domain/LottoTest.java
@@ -5,50 +5,15 @@ import static org.assertj.core.api.Assertions.assertThatCode;
 import static org.assertj.core.api.Assertions.assertThatThrownBy;
 
 import java.util.Arrays;
-import java.util.HashMap;
 import java.util.List;
 import org.junit.jupiter.api.Test;
+import org.junit.jupiter.params.ParameterizedTest;
+import org.junit.jupiter.params.provider.ValueSource;
 
 public class LottoTest {
 
-    LottoNumbers answerlottoNumbers = new LottoNumbers(
-            Arrays.asList(
-                    new SingleLottoNumber(1),
-                    new SingleLottoNumber(2),
-                    new SingleLottoNumber(3),
-                    new SingleLottoNumber(4),
-                    new SingleLottoNumber(5),
-                    new SingleLottoNumber(6)
-            )
-    );
-    SingleLottoNumber bonusNumber = new SingleLottoNumber(7);
-    Lotto answerLotto = new Lotto(answerlottoNumbers, bonusNumber);
-
     @Test
-    void 정답과_보너스_볼이_있어야_한다() {
-        assertThatCode(() -> new Lotto(answerlottoNumbers, bonusNumber)).doesNotThrowAnyException();
-    }
-
-    @Test
-    void 정답_안에_보너스_볼과_같은_번호가_포함되어_있으면_에러가_발생한다() {
-        LottoNumbers answerlottoNumbers = new LottoNumbers(
-                Arrays.asList(
-                        new SingleLottoNumber(1),
-                        new SingleLottoNumber(2),
-                        new SingleLottoNumber(3),
-                        new SingleLottoNumber(4),
-                        new SingleLottoNumber(5),
-                        new SingleLottoNumber(6)
-                )
-        );
-        SingleLottoNumber bonusNumber = new SingleLottoNumber(6);
-
-        assertThatThrownBy(() -> new Lotto(answerlottoNumbers, bonusNumber)).isInstanceOf(
-                IllegalArgumentException.class);
-    }
-
-    @Test
-    void 정답과_6개가_모두_일치하면_1등상을_받아야_한다() {
+    void 입력되는_로또_번호가_6개여야_한다() {
         List<SingleLottoNumber> singleLottoNumbers = Arrays.asList(
                 new SingleLottoNumber(1),
                 new SingleLottoNumber(2),
@@ -58,112 +23,96 @@ public class LottoTest {
                 new SingleLottoNumber(6)
         );
 
-        LottoNumbers userLotto = new LottoNumbers(singleLottoNumbers);
-        PrizeCountMap expected = new PrizeCountMap(
-                new HashMap<>() {{
-                    put(LottoPrize.FIRST_PRIZE, 1);
-                }}
-        );
-        assertThat(answerLotto.getPrizeCountMap(List.of(userLotto))).isEqualTo(expected);
+        assertThatCode(() -> new Lotto(singleLottoNumbers)).doesNotThrowAnyException();
     }
 
     @Test
-    void 정답과_5개가_일치하고_보너스_숫자와_일치하면_2등상을_받아야_한다() {
+    void 입력되는_로또_번호가_6개가_아니면_예외가_발생한다() {
         List<SingleLottoNumber> singleLottoNumbers = Arrays.asList(
                 new SingleLottoNumber(1),
                 new SingleLottoNumber(2),
                 new SingleLottoNumber(3),
                 new SingleLottoNumber(4),
                 new SingleLottoNumber(5),
+                new SingleLottoNumber(6),
                 new SingleLottoNumber(7)
         );
 
-        LottoNumbers userLotto = new LottoNumbers(singleLottoNumbers);
-        PrizeCountMap expected = new PrizeCountMap(
-                new HashMap<>() {{
-                    put(LottoPrize.SECOND_PRIZE, 1);
-                }}
-        );
-        assertThat(answerLotto.getPrizeCountMap(List.of(userLotto))).isEqualTo(expected);
+        assertThatThrownBy(() -> new Lotto((singleLottoNumbers))).isInstanceOf(IllegalArgumentException.class);
     }
 
     @Test
-    void 정답과_5개가_일치하면_3등상을_받아야_한다() {
-        List<SingleLottoNumber> singleLottoNumbers = Arrays.asList(
+    void 중복된_로또_번호가_있으면_예외가_발생한다() {
+        List<SingleLottoNumber> duplicatedLottoNumbers = Arrays.asList(
+                new SingleLottoNumber(1),
+                new SingleLottoNumber(2),
+                new SingleLottoNumber(3),
+                new SingleLottoNumber(4),
+                new SingleLottoNumber(6),
+                new SingleLottoNumber(6)
+        );
+
+        assertThatThrownBy(() -> new Lotto(duplicatedLottoNumbers)).isInstanceOf(IllegalArgumentException.class);
+    }
+
+    @ParameterizedTest
+    @ValueSource(ints = {1, 2, 3, 7, 8, 6})
+    void 보너스_번호를_받아서_포함되어_있으면_true를_반환한다(int bonusNumberValue) {
+        SingleLottoNumber bonusNumber = new SingleLottoNumber(bonusNumberValue);
+        Lotto lotto = new Lotto(
+                Arrays.asList(
+                        new SingleLottoNumber(1),
+                        new SingleLottoNumber(2),
+                        new SingleLottoNumber(3),
+                        new SingleLottoNumber(6),
+                        new SingleLottoNumber(7),
+                        new SingleLottoNumber(8)
+                )
+        );
+
+        assertThat(lotto.containsLottoNumber(bonusNumber)).isTrue();
+    }
+
+    @ParameterizedTest
+    @ValueSource(ints = {10, 11, 12, 13, 14, 15})
+    void 보너스_번호를_받아서_포함되어_있지_않으면_false를_반환한다(int bonusNumberValue) {
+        SingleLottoNumber bonusNumber = new SingleLottoNumber(bonusNumberValue);
+        Lotto lotto = new Lotto(
+                Arrays.asList(
+                        new SingleLottoNumber(1),
+                        new SingleLottoNumber(2),
+                        new SingleLottoNumber(3),
+                        new SingleLottoNumber(6),
+                        new SingleLottoNumber(7),
+                        new SingleLottoNumber(8)
+                )
+        );
+
+        assertThat(lotto.containsLottoNumber(bonusNumber)).isFalse();
+    }
+
+    @Test
+    void 로또_번호를_받아서_일치하는_번호의_개수를_반환한다() {
+        Lotto lotto = new Lotto(Arrays.asList(
                 new SingleLottoNumber(1),
                 new SingleLottoNumber(2),
                 new SingleLottoNumber(3),
                 new SingleLottoNumber(4),
                 new SingleLottoNumber(5),
-                new SingleLottoNumber(45)
+                new SingleLottoNumber(6)
+        ));
+
+        Lotto userLotto = new Lotto(
+                Arrays.asList(
+                        new SingleLottoNumber(1),
+                        new SingleLottoNumber(2),
+                        new SingleLottoNumber(3),
+                        new SingleLottoNumber(6),
+                        new SingleLottoNumber(7),
+                        new SingleLottoNumber(8)
+                )
         );
 
-        LottoNumbers userLotto = new LottoNumbers(singleLottoNumbers);
-        PrizeCountMap expected = new PrizeCountMap(
-                new HashMap<>() {{
-                    put(LottoPrize.THIRD_PRIZE, 1);
-                }}
-        );
-        assertThat(answerLotto.getPrizeCountMap(List.of(userLotto))).isEqualTo(expected);
-    }
-
-    @Test
-    void 정답과_4개가_일치하면_4등상을_받아야_한다() {
-        List<SingleLottoNumber> singleLottoNumbers = Arrays.asList(
-                new SingleLottoNumber(1),
-                new SingleLottoNumber(2),
-                new SingleLottoNumber(3),
-                new SingleLottoNumber(4),
-                new SingleLottoNumber(44),
-                new SingleLottoNumber(45)
-        );
-
-        LottoNumbers userLotto = new LottoNumbers(singleLottoNumbers);
-        PrizeCountMap expected = new PrizeCountMap(
-                new HashMap<>() {{
-                    put(LottoPrize.FOURTH_PRIZE, 1);
-                }}
-        );
-        assertThat(answerLotto.getPrizeCountMap(List.of(userLotto))).isEqualTo(expected);
-    }
-
-    @Test
-    void 정답과_3개가_일치하면_5등상을_받아야_한다() {
-        List<SingleLottoNumber> singleLottoNumbers = Arrays.asList(
-                new SingleLottoNumber(1),
-                new SingleLottoNumber(2),
-                new SingleLottoNumber(3),
-                new SingleLottoNumber(43),
-                new SingleLottoNumber(44),
-                new SingleLottoNumber(45)
-        );
-
-        LottoNumbers userLotto = new LottoNumbers(singleLottoNumbers);
-        PrizeCountMap expected = new PrizeCountMap(
-                new HashMap<>() {{
-                    put(LottoPrize.FIFTH_PRIZE, 1);
-                }}
-        );
-        assertThat(answerLotto.getPrizeCountMap(List.of(userLotto))).isEqualTo(expected);
-    }
-
-    @Test
-    void 정답과_3개_미만이_일치하면_아무것도_받지_못한다() {
-        List<SingleLottoNumber> singleLottoNumbers = Arrays.asList(
-                new SingleLottoNumber(40),
-                new SingleLottoNumber(41),
-                new SingleLottoNumber(42),
-                new SingleLottoNumber(43),
-                new SingleLottoNumber(44),
-                new SingleLottoNumber(45)
-        );
-
-        LottoNumbers userLotto = new LottoNumbers(singleLottoNumbers);
-        PrizeCountMap expected = new PrizeCountMap(
-                new HashMap<>() {{
-                    put(LottoPrize.NONE, 1);
-                }}
-        );
-        assertThat(answerLotto.getPrizeCountMap(List.of(userLotto))).isEqualTo(expected);
+        assertThat(lotto.countMatchNumber(userLotto)).isEqualTo(4);
     }
 }
